shopDjangoReact/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.5
django-debug-toolbar==3.2.1
django-grappelli==2.15.1
django-js-asset==1.2.2
django-mptt==0.12.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
phonenumbers==8.12.26
Pillow==8.3.0
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1

django-autoslug~=1.9.8
requests~=2.25.1
django-filter~=2.4.0